% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{ERA5_day_d2m}
\alias{ERA5_day_d2m}
\title{ERA5 daily 2m dewpoint temperature}
\format{
A climate4R grid object
}
\source{
ERA5 reanalysis data
}
\description{
A climate4R grid with daily dewpoint temperature (dewp) from ERA5.
}
\keyword{data}
