% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatStressIndexGrid.R
\name{indexShow}
\alias{indexShow}
\title{List all the 9 heat stress indices}
\usage{
indexShow()
}
\value{
Print a table on the screen with the following columns:
\itemize{
\item \strong{code}: Code of the index. This is the character string used as input value
for the argument \code{index.code} in \code{\link{heatStressIndexGrid}}
\item \strong{longname}: Long description of the index
\item \strong{indexfun}: The name of the internal function from package \pkg{\link{HeatStress}} used to calculate it
\item \strong{tas,dewp,hurs,wind,radiation}: A logical value (0/1) indicating the input variables required for index calculation
\item \strong{units}: The units of the index
}
}
\description{
Print a table with a summary of the 9 heat stress indices
}
\author{
J. Bedia (original)
}
