% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatStressIndexGrid.R
\name{heatStressIndexGrid}
\alias{heatStressIndexGrid}
\title{Heat Stress Indices in climate4R}
\usage{
heatStressIndexGrid(
  index.code,
  tas = NULL,
  dewp = NULL,
  hurs = NULL,
  wind = NULL,
  radiation = NULL,
  parallel = FALSE,
  max.ncores = 16,
  ncores = NULL
)
}
\arguments{
\item{index.code}{Character string, indicating the specific code of the heat stress index to be computed (see Details).}

\item{tas}{A climate4R dataset of air temperature (degrees C)}

\item{dewp}{A climate4R dataset of dew point temperature (degrees C)}

\item{hurs}{A climate4R dataset of relative humidity (\%)}

\item{wind}{A climate4R dataset of wind speed (m s-1)}

\item{radiation}{A climate4R dataset of radiation (W m-2)}

\item{parallel}{Logical. Enable parallel processing (default = FALSE)}

\item{max.ncores}{Integer. Maximum number of cores to use (default = 16)}

\item{ncores}{Integer. Specific number of cores to use}
}
\value{
A climate4R object with the computed index.
}
\description{
Calculation of 9 heat stress indices. The function is a wrapper of the package \pkg{HeatStress}
for its seamless integration with climate4R objects.
}
\details{
\code{\link{indexShow}} will display on screen a full list of heat stress indices, their codes and the required input
variables. The names of the internal functions calculating each index are also displayed, whose help files can aid in
the definition of index-specific arguments.
}
\examples{
\dontrun{
require(climate4R.HeatStress)

# Example 1: Heat Index (tas + hurs)
data("ERA5_day_t2m", package = "climate4R.HeatStress")
data("ERA5_day_hurs", package = "climate4R.HeatStress")
hi <- heatStressIndexGrid("hi", tas = ERA5_day_t2m, hurs = ERA5_day_hurs)

# Example 2: WBGT shade (tas + dewp)
data("ERA5_day_d2m", package = "climate4R.HeatStress")
wbgt_shade <- heatStressIndexGrid("wbgt_shade", tas = ERA5_day_t2m, dewp = ERA5_day_d2m)

# Example 3: WBGT sun (tas + dewp + wind + radiation)
data("ERA5_day_sfcwind", package = "climate4R.HeatStress")
data("ERA5_day_ssrd", package = "climate4R.HeatStress")
wbgt_sun <- heatStressIndexGrid("wbgt_sun", tas = ERA5_day_t2m, dewp = ERA5_day_d2m,
                                 wind = ERA5_day_sfcwind, radiation = ERA5_day_ssrd)
}

}
\author{
climate4R adaptation by C. Rodriguez-Rumayor
Original \pkg{HeatStress} by A. Casanueva \url{https://github.com/anacv/HeatStress}.
}
